+++ src/data/update-data.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+set -e
+
+## journal-abbr
+python src/data/journal-abbr/generate-journal-list-dot.py